ai_api/requirements.txt
absl-py==0.11.0
aiohttp==3.7.3
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.3.0
blinker==1.4
bokeh==2.2.3
brotlipy==0.7.0
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
cycler==0.10.0
Flask==1.1.2
gast==0.4.0
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
idna==2.10
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.3.0
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict==4.7.6
numpy==1.19.2
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.1.0
packaging==20.8
pandas==1.1.5
Pillow==8.0.1
pip==20.3.3
plotly==4.14.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyreadline==2.1
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
scipy==1.5.2
seaborn==0.11.1
setuptools==51.0.0.post20201207
sip==4.19.13
six==1.15.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.6.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
tornado==6.1
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
wheel==0.36.2
wincertstore==0.2
win-inet-pton==1.1.0
wrapt==1.12.1
yarl==1.5.1
zipp==3.4.0
#unit testing library
pytest==4.8.3
flask_sqlalchemy==2.4.4
flask-swagger-ui==3.36.0
pymssql==2.1.5